requirements.txt
Flask
python-dotenv
PyPDF2
langchain
faiss-cpu
langchain-google-genai
google-generativeai
